Added version 6.5.3206

+++ chromium/android_webview/java/src/org/chromium/android_webview/AwBrowserContext.java
@@ -10,12 +10,13 @@ import android.content.SharedPreferences;
 import androidx.annotation.NonNull;
 import androidx.annotation.VisibleForTesting;
 
+import org.jni_zero.CalledByNative;
+import org.jni_zero.JNINamespace;
+import org.jni_zero.NativeMethods;
+
 import org.chromium.android_webview.common.Lifetime;
 import org.chromium.base.ContextUtils;
 import org.chromium.base.StrictModeContext;
-import org.chromium.base.annotations.CalledByNative;
-import org.chromium.base.annotations.JNINamespace;
-import org.chromium.base.annotations.NativeMethods;
 import org.chromium.base.memory.MemoryPressureMonitor;
 import org.chromium.content_public.browser.BrowserContextHandle;
 import org.chromium.content_public.browser.ContentViewStatics;
@@ -71,8 +72,7 @@ public class AwBrowserContext implements BrowserContextHandle {
 
         try (StrictModeContext ignored = StrictModeContext.allowDiskWrites()) {
             // Prefs dir will be created if it doesn't exist, so must allow writes.
-            mSharedPreferences = ContextUtils.getApplicationContext().getSharedPreferences(
-                    getSharedPrefsFilename(relativePath), Context.MODE_PRIVATE);
+            mSharedPreferences = createSharedPrefs(relativePath);
 
             if (isDefaultAwBrowserContext()) {
                 // Migration requires disk writes.
@@ -83,16 +83,17 @@ public class AwBrowserContext implements BrowserContextHandle {
         // Register MemoryPressureMonitor callbacks and make sure it polls only if there is at
         // least one WebView around.
         MemoryPressureMonitor.INSTANCE.registerComponentCallbacks();
-        AwContentsLifecycleNotifier.addObserver(new AwContentsLifecycleNotifier.Observer() {
-            @Override
-            public void onFirstWebViewCreated() {
-                MemoryPressureMonitor.INSTANCE.enablePolling();
-            }
-            @Override
-            public void onLastWebViewDestroyed() {
-                MemoryPressureMonitor.INSTANCE.disablePolling();
-            }
-        });
+        AwContentsLifecycleNotifier.getInstance().addObserver(
+                new AwContentsLifecycleNotifier.Observer() {
+                    @Override
+                    public void onFirstWebViewCreated() {
+                        MemoryPressureMonitor.INSTANCE.enablePolling();
+                    }
+                    @Override
+                    public void onLastWebViewDestroyed() {
+                        MemoryPressureMonitor.INSTANCE.disablePolling();
+                    }
+                });
     }
 
     @VisibleForTesting
@@ -232,10 +233,15 @@ public class AwBrowserContext implements BrowserContextHandle {
      * <p>
      * Name must be non-null and valid Unicode.
      *
-     * @throws IllegalStateException if trying to delete the default profile or a profile which is
-     *                               in use.
+     * @throws IllegalArgumentException if trying to delete the default profile.
+     * @throws IllegalStateException if trying to delete a profile which is in use.
      */
-    public static boolean deleteNamedContext(String name) {
+    public static boolean deleteNamedContext(String name)
+            throws IllegalArgumentException, IllegalStateException {
+        final String defaultContextName = AwBrowserContextJni.get().getDefaultContextName();
+        if (name.equals(defaultContextName)) {
+            throw new IllegalArgumentException("Cannot delete the default profile");
+        }
         return AwBrowserContextJni.get().deleteNamedContext(name);
     }
 
@@ -273,11 +279,30 @@ public class AwBrowserContext implements BrowserContextHandle {
         AwBrowserContextJni.get().clearFormData(mNativeAwBrowserContext);
     }
 
+    public void setServiceWorkerIoThreadClient(AwContentsIoThreadClient ioThreadClient) {
+        AwBrowserContextJni.get().setServiceWorkerIoThreadClient(
+                mNativeAwBrowserContext, ioThreadClient);
+    }
+
+    private static SharedPreferences createSharedPrefs(String relativePath) {
+        return ContextUtils.getApplicationContext().getSharedPreferences(
+                getSharedPrefsFilename(relativePath), Context.MODE_PRIVATE);
+    }
+
     @CalledByNative
     public static AwBrowserContext create(long nativeAwBrowserContext, String name,
             String relativePath, AwCookieManager cookieManager, boolean isDefault) {
         return new AwBrowserContext(
                 nativeAwBrowserContext, name, relativePath, cookieManager, isDefault);
+    }
+
+    @CalledByNative
+    public static void deleteSharedPreferences(String relativePath) {
+        try (StrictModeContext ignored = StrictModeContext.allowDiskWrites()) {
+            final String sharedPrefsFilename = getSharedPrefsFilename(relativePath);
+            SharedPreferences.Editor prefsEditor = createSharedPrefs(sharedPrefsFilename).edit();
+            prefsEditor.clear().apply();
+        }
     }
 
     @NativeMethods
@@ -297,5 +322,7 @@ public class AwBrowserContext implements BrowserContextHandle {
         void clearPersistentOriginTrialStorageForTesting(long nativeAwBrowserContext);
         boolean hasFormData(long nativeAwBrowserContext);
         void clearFormData(long nativeAwBrowserContext);
+        void setServiceWorkerIoThreadClient(
+                long nativeAwBrowserContext, AwContentsIoThreadClient ioThreadClient);
     }
 }
